Prevent Signature swapping within our APIs
Since the updated Schnorr API makes use of the same Signature class as ECDSA, we need to make sure signature swapping at our API layer is caught and rejected.

This isn't a cryptographic mitigation. It is not meant to guard against algorithm confusion attacks.

This is only intended to prevent a trivial kind of developer mistake.